requirements.txt
asgiref==3.8.1
async-timeout==5.0.1
Django==4.2.19
django-redis==5.4.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
mysqlclient==2.2.7
PyJWT==2.10.1
python-dotenv==1.0.1
redis==5.2.1
sqlparse==0.5.3
typing_extensions==4.12.2
